refactor(asset): unecessary cast
`status` is already a `u16`, so it does not need to be casted again.

+++ crates/bevy_asset/src/io/wasm.rs
@@ -81,7 +81,7 @@ impl HttpWasmAssetReader {
                 Ok(reader)
             }
             404 => Err(AssetReaderError::NotFound(path)),
-            status => Err(AssetReaderError::HttpError(status as u16)),
+            status => Err(AssetReaderError::HttpError(status)),
         }
     }
 }
